Changed QualtyValueStrategy implemetations from 1 abstract class, 2 subclasses to 1 enum with 2 values.  fixed bug in DefaultContig.Builder
git-svn-id: svn+ssh://svn.code.sf.net/p/jillion/code/trunk@365 d1f56cec-0f3b-4699-95c7-aedf87c10f2a

+++ src-internal/org/jcvi/assembly/analysis/InternalSangerContigChecker.java
@@ -26,7 +26,7 @@ public class InternalSangerContigChecker<R extends PlacedRead> extends ContigChe
     protected List<Process> createProcesses() {
         List<Process> analyizers = new ArrayList<Process>();
         analyizers.addAll(super.createProcesses());
-        analyizers.add(new QualityClassContigMapAnalysisProcess<R>(this.getContigCheckBuilder(), this.getStruct(),QualityClass.valueOf((byte)10)));
+        analyizers.add(new QualityClassContigMapAnalysisProcess<R>(this.getContigCheckBuilder(), this.getStruct(),QualityClass.valueOf(10)));
         return analyizers;
     }
